Script `lts:patch` uses symfony/console (added to `zf-maintainer`)
