+++ bin/zf-maintainer
@@ -19,6 +19,7 @@ $application->add(new \ZF\Maintainer\CreatePackage());
 $application->add(new \ZF\Maintainer\RebaseDocTemplates());
 
 $application->add((new \ZF\Maintainer\Lts\Components())->setComponents($components));
+$application->add(new \ZF\Maintainer\Lts\Patch());
 $application->add((new \ZF\Maintainer\Lts\Release())->setComponents($components));
 
 $application->run();
